+++ packages/assets-controllers/src/TokenDetectionController.test.ts
@@ -1193,6 +1193,7 @@ describe('TokenDetectionController', () => {
                 '0xa',
                 '0x89',
                 '0x531',
+                '0x279f',
               ],
               selectedAddress: secondSelectedAccount.address,
             });